+++ CHANGELOG.md
@@ -44,7 +44,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Specify dtype for empty pandas.Series for forward compatibility
   - Remove wildcard imports, relative imports
   - Include flake8 check as part of testing suites
-  - Improve unit testing coverage of instrument functions
+  - Improve unit testing coverage of instrument functions and instrument object
 
 ## [2.2.0] - 2020-07-06
 - New Features